Creating the NavigationView and the TabView
