--- Twitter clone/Twitter clone/ContentView.swift
@@ -1,21 +1,0 @@
-//
-//  ContentView.swift
-//  Twitter clone
-//
-//  Created by macbook on 1/23/21.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        Text("Hello, world!")
-            .padding()
-    }
-}
-
-struct ContentView_Previews: PreviewProvider {
-    static var previews: some View {
-        ContentView()
-    }
-}
