+++ Twitter clone/Views/ContentView.swift
@@ -1,0 +1,41 @@
+//
+//  ContentView.swift
+//  Twitter clone
+//
+//  Created by macbook on 1/23/21.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    var body: some View {
+        NavigationView {
+            TabView {
+                Text("Feed")
+                    .tabItem {
+                        Image(systemName: "house")
+                        Text("Home")
+                    }
+                Text("Search")
+                    .tabItem {
+                        Image(systemName: "magnifyingglass")
+                        Text("Search")
+                    }
+                Text("Messages")
+                    .tabItem {
+                        Image(systemName: "envelope")
+                        Text("Messages")
+                    }
+            }
+            // Bar title
+            .navigationBarTitle("Home")
+            .navigationBarTitleDisplayMode(.inline)
+        }
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
